users/login-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
python-jose
email-validator
passlib
python-dotenv